(REF) Rename WorkflowMessageExample => ExampleData
